requirements.txt
Django==4.0
djangorestframework==3.12.4
djangorestframework-jsonapi
djoser
pytz
djangorestframework-jwt